pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "factored-medical-qa-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.11"
dependencies = [
    "datasets>=4.0.0",
    "faiss-cpu>=1.7.4",
    "ipykernel>=6.30.0",
    "nltk>=3.9.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "transformers>=4.53.3",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "accelerate>=0.26.0",
    "streamlit>=1.36.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.pip]
extra-index-url = ["https://download.pytorch.org/whl/cu118"]
